requirements.txt
openai>=1.0.0
pinecone-client>=2.0.0
chromadb>=0.4.0
qdrant-client>=1.7.0
weaviate-client>=4.0.0
python-dotenv>=1.0.0
